v46.9 - The Nice Update
changed keybinds for gadgets: for frag grenade from 'g' to 'f', and for sticky grenade from 'v' to 'r'. impulse is still 'e' (because unlike everything else, impulse has always been perfect from the very start. i'm proud of you, impulse grenade).

bumped this version from v46.8 (which was hundreds of big and small changes) to v46.9 (which is technically only 1 change) because i did it by hand, and you won't believe how many hard it is to look for a single letter that shows up multiple times and is spread out everywhere, much less TWO of them. it only makes sense.

also, this is a nice update. you're welcome. ;)