v46.9.1R - Tiny Display Text Revisions
this is still just v46.9.1 but with corrected display texts on the start screen--since v46.9 the frag and sticky grenades have been remapped to 'f' and 'r' keys but the instruction still says 'hold g/v/e to aim gadget', so that's corrected now.

also removed the "(Nice)" wording from display game titles (i was being a little silly). but hey, still a nice little update!